+++ sdk/src/main/java/ly/count/android/sdk/ModuleAttribution.java
@@ -11,6 +11,12 @@ public class ModuleAttribution extends ModuleBase {
         attributionInterface = new ModuleAttribution.Attribution();
     }
 
+    void recordCampaignInternal(String campaignId, String campaignUserId) {
+        L.d("[ModuleAttribution] recordCampaignInternal, campaign id:[" + campaignId + "], user id:[" + campaignUserId + "]");
+
+        _cly.connectionQueue_.sendReferrerDataManual(campaignId, campaignUserId);
+    }
+
     @Override
     public void halt() {
         attributionInterface = null;
@@ -18,12 +24,28 @@ public class ModuleAttribution extends ModuleBase {
 
     public class Attribution {
 
+        /**
+         * Report user attribution manually
+         * @param campaignId
+         */
         public void recordCampaign(String campaignId) {
+            synchronized (_cly) {
+                L.i("[Attribution] calling 'recordCampaign'");
 
+                recordCampaignInternal(campaignId, null);
+            }
         }
 
+        /**
+         * Report user attribution manually
+         * @param campaignId
+         */
         public void recordCampaign(String campaignId, String campaignUserId) {
+            synchronized (_cly) {
+                L.i("[Attribution] calling 'recordCampaign'");
 
+                recordCampaignInternal(campaignId, campaignUserId);
+            }
         }
     }
 }